fixed test cases for ruby 2.4
