+++ lib/glossarist/designation/grammar_info.rb
@@ -7,6 +7,7 @@ module Glossarist
     class GrammarInfo
       include Glossarist::Utilities::Enum
       include Glossarist::Utilities::BooleanAttributes
+      include Glossarist::Utilities::CommonFunctions
 
       register_enum :gender, Glossarist::GlossaryDefinition::GRAMMAR_INFO_GENDERS, multiple: true
       register_enum :number, Glossarist::GlossaryDefinition::GRAMMAR_INFO_NUMBERS, multiple: true
@@ -51,28 +52,6 @@ module Glossarist
           :noun,
           :part_of_speech,
         ])
-      end
-
-      # Hash#transform_keys is not available in Ruby 2.4
-      # so we have to do this ourselves :(
-      # symbolize hash keys
-      def symbolize_keys(hash)
-        result = {}
-        hash.each_pair do |key, value|
-          result[key.to_sym] = value
-        end
-        result
-      end
-
-      # Hash#slice is not available in Ruby 2.4
-      # so we have to do this ourselves :(
-      # slice hash keys
-      def slice_keys(hash, keys)
-        result = {}
-        keys.each do |key|
-          result[key] = hash[key] if hash.key?(key)
-        end
-        result
       end
     end
   end
