+++ lib/glossarist/utilities.rb
@@ -2,3 +2,4 @@
 
 require_relative "utilities/enum"
 require_relative "utilities/boolean_attributes"
+require_relative "utilities/common_functions"
